Move records into modules
