+++ src/lib/index.ts
@@ -7,11 +7,9 @@ export { OutboundTransporter } from './transport/OutboundTransporter';
 export { encodeInvitationToUrl, decodeInvitationFromUrl } from './helpers';
 export { InitConfig, OutboundPackage } from './types';
 
-export { ConnectionRecord } from './storage/ConnectionRecord';
-export { BasicMessageRecord } from './storage/BasicMessageRecord';
-export { CredentialRecord } from './storage/CredentialRecord';
-export { ConnectionEventType } from './modules/connections/services/ConnectionService';
-export { BasicMessageEventType } from './modules/basic-messages/services/BasicMessageService';
+export { ConnectionRecord, ConnectionEventType } from './modules/connections';
+export { BasicMessageRecord, BasicMessageEventType } from './modules/basic-messages';
+export { CredentialRecord } from './modules/credentials';
 export * from './modules/credentials';
 export * from './modules/proofs';
 export * from './modules/connections';
