+++ src/lib/modules/basic-messages/BasicMessagesModule.ts
@@ -2,7 +2,7 @@ import type { WalletQuery } from 'indy-sdk';
 import { EventEmitter } from 'events';
 import { BasicMessageService } from './services';
 import { MessageSender } from '../../agent/MessageSender';
-import { ConnectionRecord } from '../../storage/ConnectionRecord';
+import { ConnectionRecord } from '../connections';
 import { Dispatcher } from '../../agent/Dispatcher';
 import { BasicMessageHandler } from './handlers/BasicMessageHandler';
