+++ src/lib/modules/basic-messages/index.ts
@@ -1,2 +1,3 @@
 export * from './messages';
 export * from './services';
+export * from './repository/BasicMessageRecord';
