+++ src/lib/modules/connections/index.ts
@@ -1,3 +1,4 @@
 export * from './messages';
 export * from './services';
 export * from './models';
+export * from './repository/ConnectionRecord';
