+++ src/lib/modules/credentials/index.ts
@@ -2,3 +2,4 @@ export * from './messages';
 export * from './services';
 export * from './CredentialUtils';
 export * from './models';
+export * from './repository/CredentialRecord';
