+++ src/lib/modules/proofs/ProofsModule.ts
@@ -2,7 +2,7 @@ import { createOutboundMessage } from '../../agent/helpers';
 import { MessageSender } from '../../agent/MessageSender';
 import { ConnectionService } from '../connections';
 import { ProofService } from './services';
-import { ProofRecord } from '../../storage/ProofRecord';
+import { ProofRecord } from './repository/ProofRecord';
 import { ProofRequest } from './models/ProofRequest';
 import { JsonTransformer } from '../../utils/JsonTransformer';
 import { EventEmitter } from 'events';
