+++ src/lib/modules/proofs/index.ts
@@ -2,3 +2,4 @@ export * from './messages';
 export * from './models';
 export * from './services';
 export * from './ProofState';
+export * from './repository/ProofRecord';
